% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flowFrame.R
\name{as.flowFrame,fcbFlowFrame-method}
\alias{as.flowFrame,fcbFlowFrame-method}
\title{Coerces fcbFlowFrames to flowFrames by extracting the barcoded.ff}
\usage{
\S4method{as.flowFrame}{fcbFlowFrame}(x, ...)
}
\arguments{
\item{fcbFlowFrame}{fcbFlowFrame object with barcoded flowframe and uptake flowframe post deskewing,
clustering (at least one barcodes slot filled), and assigning}
}
\value{
the barcoded flowframe from slot 1
}
\description{
Coerces fcbFlowFrames to flowFrames by extracting the barcoded.ff
}
