% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flowSet.R
\name{as.flowSet}
\alias{as.flowSet}
\title{Writes FCS files for each well/level and returns debarcoded flowSet}
\usage{
as.flowSet(flowFrameFCB)
}
\arguments{
\item{flowFrameFCB}{flowFrameFCB object with barcoded flowframe and uptake flowframe post deskewing,
clustering (at least one barcodes slot filled), and assigning}

\item{simplify}{if TRUE, return only the debarcoded flowSet; if FALSE, return
debarcoded flowSet as slot in flowFrameFCB (default is TRUE)}
}
\value{
a matrix of probabilities, with ncol = levels, and nrow = legnth(vec).
}
\description{
Writes FCS files for each well/level and returns debarcoded flowSet
}
