% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_fcbFlowSet.R
\name{assign_fcbFlowSet}
\alias{assign_fcbFlowSet}
\title{Assign cells to barcoding levels for a fcbFlowSet}
\usage{
assign_fcbFlowSet(fcbFlowSet, channel, likelihoodcut = 8, ambiguitycut = 0.02)
}
\arguments{
\item{fcbFlowSet}{An fcbFlowSet object with deskewing and clustering completed.}

\item{channel}{The name (string) of the channel that has been corrected and clustered.}

\item{likelihoodcut}{Numeric, a likelihood cutoff for discarding unlikely cells.
Cells less than 1/k as likely as the most likely cell from that population are unassigned.}

\item{ambiguitycut}{Numeric from 0 to 1, threshold below which to discard ambiguous cells.
E.g., 0.02 discards cells with more than 2\% chance of originating from another population.}
}
\value{
An fcbFlowSet object with assignment slots populated.
}
\description{
Assign cells to barcoding levels for a fcbFlowSet
}
