% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_flowFrameFCB.R
\name{assign_flowFrameFCB}
\alias{assign_flowFrameFCB}
\title{Defines populations on barcoded datasets}
\usage{
assign_flowFrameFCB(
  flowFrameFCB,
  channel,
  likelihoodcut = 8,
  ambiguitycut = 0.02
)
}
\arguments{
\item{channel}{string, channel which is barcoded, only needed if output == "plot"}

\item{likelihoodcut}{numeric, a likelihood cutoff for discarding unlikely cells, less than 1/k as likely as the most likely cell from that population}

\item{ambiguitycut}{numeric from 0 to 1, threshhold below which to discard ambigious cells, eg: 0.02,
discards cells with more than 2% chance of originating from another population}

\item{fcb_df}{dataframe, representing the barcoded data, columns for each channel, rows for each cell}

\item{probs}{matrix of probabilites under population models, returned by 'fit_models'}

\item{output}{string, default "classif," also "plot;" what to return?}
}
\value{
a vector of integers from 0:ncol(probs), cells assigned a classification of 0 remained unassigned,
 otherwise number corresponds to the barcoding level assignment of that cell
}
\description{
Defines populations on barcoded datasets
}
