% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_flowFrameFCB.R
\name{cluster_flowFrameFCB}
\alias{cluster_flowFrameFCB}
\title{Defines populations on barcoded datasets}
\usage{
cluster_flowFrameFCB(
  flowFrameFCB,
  channel,
  levels,
  opt = "mixture",
  dist = NULL,
  subsample = 10000,
  trim = 0,
  updateProgress = NULL
)
}
\arguments{
\item{levels}{integer, the number of barcoding intensities present in the vector}

\item{opt}{string, either "mixture" (default) for gaussian mixture modeling, or "fisher" for fisher-jenks natural breaks optimization}

\item{dist}{string in c("normal, skew.Normal, Tdist"), passed to mixsmsn}

\item{subsample}{Integer, number of cells to subsample, defaults to 10,000}

\item{trim}{numberic between 0, 1; used to trim the upper and lower extremes to exlcude outliers (eg. trim = 0.01 exludes most extreme 1% of data)}

\item{updateProgress}{used in reactive context (shiny) to return progress information to GUI}

\item{vec}{a vector of barcode intensities, usually post morphology correction}
}
\value{
a matrix of probabilities, with ncol = levels, and nrow = legnth(vec).
If gaussian mixture modeling is used the probailities correspond to the probability
of the cell originaiting that level under the distrubtion specified by the mixture model
If jenks natural breaks optimization is used, the probability is estimated empirically based on a histogram
}
\description{
This function allows you to calculate the probability of a cell originating from a given population using
either gaussian mixture modeling or jenks natural breaks classification
}
\seealso{
\code{\link{deskew_flowFrameFCB}}
}
