% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRegressContrained.R
\name{constrained_regression}
\alias{constrained_regression}
\title{Constrained regression for morphology correction}
\usage{
constrained_regression(
  X,
  Y,
  fsc_limits,
  ssc_limits,
  val3,
  D,
  OFFSET,
  Bx,
  rm,
  monodir
)
}
\arguments{
\item{X}{Design matrix for regression.}

\item{Y}{Response vector.}

\item{fsc_limits}{Forward scatter limits.}

\item{ssc_limits}{Side scatter limits.}

\item{val3}{Additional parameter for regression.}

\item{D}{Data matrix for density estimation.}

\item{OFFSET}{Offset parameter.}

\item{Bx}{Basis expansion matrix.}

\item{rm}{Regression model matrix.}

\item{monodir}{Monotonicity direction constraints.}
}
\value{
Constrained regression result.
}
\description{
Internal function used to build the regression model with monotonicity
constraints. Not intended to be called directly by the user.
}
\seealso{
\code{\link{doRegressConstrained}}
}
