% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deskew_flowFrameFCB.R
\name{deskew_flowFrameFCB}
\alias{deskew_flowFrameFCB}
\title{Corrects morphology based on scatter or uptake control for flowFrameFCB}
\usage{
deskew_flowFrameFCB(
  flowFrameFCB,
  channel,
  method = c("earth", "knignenburg", "lm"),
  predictors = c("fsc_a", "ssc_a"),
  subsample = 30000,
  ret.model = TRUE,
  verbose = FALSE,
  updateProgress = NULL,
  ...
)
}
\arguments{
\item{channel}{The name (string) of the channel to be corrected, ie. the column name in 'fcb_df'}

\item{predictors}{The vector of channel names to be used to build the regression model}

\item{subsample}{Integer, number of cells to sample (with replacement) for the morphology correction, defaults to 10,000.}

\item{updateProgress}{used in reactive context (shiny) to return progress information to GUI#'}

\item{fcb}{the barcoded dataframe, post compensation and preprocessing}

\item{uptake}{Optional: a dataframe consisting of all cells barcoded with a single level of the barcoding dye}
}
\value{
a tibble/data.frame with the selected channel corrected for fsc and ssc
}
\description{
Corrects morphology based on scatter or uptake control for flowFrameFCB
}
