% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRegressContrained.R
\name{doRegressContrained}
\alias{doRegressContrained}
\title{obtain residulas afer monitonically constrained regression}
\usage{
doRegressContrained(single_level_bc, fcb_df = NULL, Loc, weight, trans,
  fsc_ssc = c(fsc = "FSC-A", ssc = "SSC-A"), val3 = NULL,
  constrained_flag = 1, columns = NULL, monodir = NULL,
  cofactor = NULL)
}
\arguments{
\item{single_level_bc}{text}

\item{fcb_df}{text}

\item{Loc}{text}

\item{trans}{text}

\item{val3}{text}

\item{constrained_flag}{text}

\item{columns}{text}

\item{monodir}{text}

\item{cofactor}{text}

\item{Weight}{text}
}
\value{
some text
}
\description{
obtain residulas afer monitonically constrained regression
}
\examples{
text
}
\seealso{
text
}
