% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcbFlowFrame.R
\name{fcbFlowFrame-class}
\alias{fcbFlowFrame-class}
\alias{fcbFlowFrame}
\title{fcbFlowFrame class}
\usage{
fcbFlowFrame(x, barcodes = list())

fcbFlowFrame(x, barcodes = list())
}
\arguments{
\item{x}{A \code{flowFrame} or \code{cytoframe} object.}

\item{barcodes}{A list of barcode data (default: empty list).}
}
\value{
An object of class \code{fcbFlowFrame}.
}
\description{
A container for barcoded flow cytometry data, extending flowCore's
\code{flowFrame} class with a \code{barcodes} slot that stores
deskewing, clustering, and assignment results.

Constructor function to create an fcbFlowFrame from a flowFrame or
cytoframe. If a \code{cytoframe} is provided (from the \code{flowWorkspace}
package), it is automatically converted to a \code{flowFrame} first.
}
\section{Slots}{

\describe{
\item{\code{barcodes}}{A named list storing debarcoding results per channel.
Each element is named by channel and contains sublists for
\code{deskewing}, \code{clustering}, and \code{assignment}.}
}}

\examples{
data(jurkatFCB)
fcb <- fcbFlowFrame(jurkatFCB)
fcb
}
\seealso{
\code{\link{deskew_fcbFlowFrame}} to begin the debarcoding pipeline,
  \code{\link{as.flowFrame}} and \code{\link{as.cytoframe}} for coercion
}
