% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcbFlowSet.R
\docType{class}
\name{fcbFlowSet-class}
\alias{fcbFlowSet-class}
\alias{.fcbFlowSet}
\title{fcbFlowSet class}
\description{
A container for debarcoded flow cytometry data, extending flowCore's
\code{flowSet} class. Contains a collection of flowFrames, one per
debarcoded sample.
}
\examples{
data(jurkatFCB)
library(flowCore)
fcbfs <- fcbFlowSet(flowSet(list(sample1 = jurkatFCB)))
class(fcbfs)
}
