% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcbFlowSet.R
\name{fcbFlowSet}
\alias{fcbFlowSet}
\title{Create an fcbFlowSet}
\usage{
fcbFlowSet(x)
}
\arguments{
\item{x}{A \code{flowSet} or \code{cytoset} object.}
}
\value{
An object of class \code{fcbFlowSet}.
}
\description{
Constructor function to create an fcbFlowSet from a flowSet or cytoset.
If a \code{cytoset} is provided (from the \code{flowWorkspace} package),
it is automatically converted to a \code{flowSet} first.
}
\examples{
\dontrun{
# fcbFlowSet is normally created via split(); direct construction is also
# possible if you already have a flowSet:
data(jurkatFCB)
library(flowCore)
fs <- flowSet(list(sample1 = jurkatFCB))
fcbfs <- fcbFlowSet(fs)
}
}
\seealso{
\code{\link{split}} to produce an fcbFlowSet from an assigned
  fcbFlowFrame, \code{\link{as.cytoset}} for coercion back to cytoset
}
