% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Defines populations on barcoded datasets}
\usage{
fit_models(vec, levels, opt = "mixture", dist = NULL,
  trans = "arcsinh", cofactor = NULL, subsample = 10000,
  updateProgress = NULL)
}
\arguments{
\item{vec}{a vector of barcode intensities, usually post morphology correction}

\item{levels}{integer, the number of barcoding intensities present in the vector}

\item{opt}{string, either "mixture" (default) for gaussian mixture modeling, or "fisher" for fisher-jenks natural breaks optimization}

\item{dist}{string in c("normal, skew.Normal, Tdist"), passed to mixsmsn}

\item{trans}{string, transformation to apply to the barcoding channels, defaults to arcsinh, also log10}

\item{cofactor}{Numeirc, Cofactor used for the arcsinh transformation on vec}

\item{subsample}{Integer, number of cells to subsample, defaults to 10,000}

\item{updateProgress}{used in reactive context (shiny) to return progress information to GUI}
}
\value{
a matrix of probabilities, with ncol = levels, and nrow = legnth(vec).
If gaussian mixture modeling is used the probailities correspond to the probability
of the cell originaiting that level under the distrubtion specified by the mixture model
If jenks natural breaks optimization is used, the probability is estimated empirically based on a histogram
}
\description{
This function allows you to calculate the probability of a cell originating from a given population using
either gaussian mixture modeling or jenks natural breaks classificatio.n
}
\examples{

}
\seealso{
\code{\link{morphology_corr}}
}
