% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deskew_fcbFlowFrame.R,
%   R/morphology_corr_master.R
\name{match.arg1}
\alias{match.arg1}
\title{Option selector.}
\usage{
match.arg1(arg, choices)

match.arg1(arg, choices)
}
\arguments{
\item{arg}{a choice or a vector of choices}

\item{choices}{a character vector of choices}
}
\value{
returns the first choice is no choice is made, otherwise returns
 a the choice, or an error if the choice was invaldi

returns the first choice is no choice is made, otherwise returns
 a the choice, or an error if the choice was invaldi
}
\description{
Option selector.

Option selector.
}
