% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology_corr.R
\name{morphology_corr_old}
\alias{morphology_corr_old}
\title{Corrects morphology based on scatter or uptake control.}
\usage{
morphology_corr_old(
  fcb_df,
  bc_single_level = NULL,
  channel,
  opt = "regression",
  subsample = 10000,
  trans = "arcsinh",
  updateProgress = NULL,
  cofactor_bc1 = NULL,
  cofactor_uptake = NULL,
  uptake_channel = NULL
)
}
\arguments{
\item{fcb_df}{the barcoded dataframe, post compensation and preprocessing}

\item{bc_single_level}{Optional: a dataframe consisting of all cells barcoded with a single level of the barcoding dye}

\item{channel}{The name (string) of the channel to be corrected, ie. the column name in 'fcb_df'}

\item{opt}{String, Morphology correction apporach, either "regression" (default) or "control dye"}

\item{subsample}{Integer, number of cells to sample (with replacement) for the morphology correction, defaults to 10,000.}

\item{trans}{String, transformation to apply to the barcoding channels, defaults to arcsinh, also log10}

\item{updateProgress}{used in reactive context (shiny) to return progress information to GUI}

\item{cofactor_bc1}{Numeirc, Cofactor used for the arcsinh transformation on the barcoding channels}

\item{uptake_channel}{String, uptake channel if "control dye" option is used as morphology correction method}

\item{cofactor_update}{Numeric, Cofactor used for the arcsinh transformation of the uptake channel}
}
\value{
numeric vector of length nrow(fcb_df) representing the morphology corrected channel
}
\description{
Corrects morphology based on scatter or uptake control.
}
