% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fcbFlowFrame-assignments.R
\name{plot.fcbflowframe}
\alias{plot.fcbflowframe}
\title{Plot debarcoding assignments}
\usage{
plot.fcbflowframe(fcbFlowFrame, plot = "assignments", seed = 1)
}
\arguments{
\item{fcbFlowFrame}{An fcbFlowFrame with completed assignments.}

\item{plot}{Character, plot type: "assignments" (default), "density", "chull", or "data".}

\item{seed}{Integer, random seed for color palette generation (default 1).}
}
\description{
Plot debarcoding assignments
}
