% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDenseScatterArea.R
\name{selectDenseScatterArea}
\alias{selectDenseScatterArea}
\title{Select dense scatter area}
\usage{
selectDenseScatterArea(
  data,
  fsc_ssc = c(fsc = "FSC-A", ssc = "SSC-A"),
  subsample = 10000
)
}
\arguments{
\item{data}{Data frame or matrix of flow cytometry data.}

\item{fsc_ssc}{Named character vector with FSC and SSC channel names.}

\item{subsample}{Integer, number of cells to subsample for density estimation.}
}
\description{
Finds the area containing 95\% of cells based on FSC/SSC density,
used for the Knijnenburg morphology correction method.
}
