% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_platemap.R
\name{apply_platemap}
\alias{apply_platemap}
\title{Updates pData and filenames with well assignments from platemap}
\usage{
apply_platemap(fcbFlowSet, platemap, drop0 = FALSE, prefix = NA)
}
\arguments{
\item{fcbFlowSet}{debarcoded flowframe(s)}

\item{platemap}{data.frame, with the well assignments}

\item{drop0}{logical, drop unassigned events, default = FALSE}

\item{prefix}{character, file prefix for new FCS files, if not specifed, will pull from FCS file $FIL keyword}
}
\value{
an fcbFlowSet object with updated names and pData
}
\description{
Updates pData and filenames with well assignments from platemap
}
