% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flowFrame.R
\name{as.flowFrame}
\alias{as.flowFrame}
\title{Generic function for coercing flowFrame like objects to flowFrame}
\usage{
as.flowFrame(x, ...)
}
\arguments{
\item{x}{flowFrame like object}

\item{...}{other parameters passed to specific methods}
}
\description{
Generic function for coercing flowFrame like objects to flowFrame
}
