% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_fcbFlowFrame.R
\name{calculate.ambiguity}
\alias{calculate.ambiguity}
\title{Ambiguity cutoff}
\usage{
calculate.ambiguity(probs)
}
\arguments{
\item{fcbFlowFrame}{a fcbFlowFrame object with barcoded flowframe and uptake flowframe post deskewing
and clustering (at least one barcodes slot filled)}
}
\value{
the prbability matrix normalized by row
}
\description{
Ambiguity cutoff
}
