% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_fcbFlowFrame.R,
%   R/assign_flowFrameFCBOld.R
\name{calculate.likelihood}
\alias{calculate.likelihood}
\title{Likelihood cutoff}
\usage{
calculate.likelihood(fcbFlowFrame, channel)

calculate.likelihood(fcbFlowFrame, channel)
}
\arguments{
\item{fcbFlowFrame}{a fcbFlowFrame object with barcoded flowframe and uptake flowframe post deskewing
and clustering (at least one barcodes slot filled)}
}
\value{
the prbability matrix normalized by column

the prbability matrix normalized by column
}
\description{
Likelihood cutoff

Likelihood cutoff
}
