% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fcbFlowFrame.R
\name{cluster_fcbFlowFrame}
\alias{cluster_fcbFlowFrame}
\title{Defines populations on barcoded datasets}
\usage{
cluster_fcbFlowFrame(
  fcbFlowFrame,
  channel,
  levels,
  opt = "mixture",
  dist = NULL,
  subsample = 3000,
  trim = 0,
  ret.model = TRUE,
  manbreaks = NULL,
  updateProgress = NULL
)
}
\arguments{
\item{fcbFlowFrame}{a fcbFlowFrame object with barcoded flowframe and uptake flowframe post deskewing (at least one barcodes slot filled)}

\item{channel}{The name (string) of the channel to be clustered}

\item{levels}{integer, the number of barcoding intensities present in the vector}

\item{opt}{string, either "mixture" (default) for gaussian mixture modeling, or "fisher" for fisher-jenks natural breaks optimization, or "manual.breaks" for manual specification of breakpoints}

\item{dist}{string in c("Normal, Skew.normal, Tdist"), passed to mixsmsn}

\item{subsample}{Integer, number of cells to subsample, defaults to 10,000}

\item{trim}{numberic between 0, 1; used to trim the upper and lower extremes to exlcude outliers (eg. trim = 0.01 exludes most extreme 1\% of data)}

\item{ret.model}{Option to retain the model for deskewing}

\item{manbreaks}{Vector, length levels + 1, used to specifiy manual breakpoints for levels}

\item{updateProgress}{used in reactive context (shiny) to return progress information to GUI}
}
\value{
a fcbFlowFrame with deskewed barcodes slot and clustering slot with a matrix of probabilities, with ncol = levels, and nrow = legnth(vec).
If gaussian mixture modeling is used the probailities correspond to the probability
of the cell originaiting that level under the distrubtion specified by the mixture model
If jenks natural breaks optimization is used, the probability is estimated empirically based on a histogram
}
\description{
This function allows you to calculate the probability of a cell originating from a given population using
either gaussian mixture modeling or jenks natural breaks classification
}
\seealso{
\code{\link{deskew_fcbFlowFrame}}
}
