% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRegressContrained.R
\name{constrained_regression}
\alias{constrained_regression}
\title{Used to build the regression model, not to be called directly by the used}
\usage{
constrained_regression(X, Y, fsc_limits, ssc_limits, val3, D, OFFSET, Bx,
  rm, monodir)
}
\value{
numeric vector of length nrow(fcb_df) representing the morphology corrected channel
}
\description{
Used to build the regression model, not to be called directly by the used
}
\seealso{
\code{\link{doRegressConstrained}}
}
