% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deskew_fcbFlowFrame.R
\name{deskew_fcbFlowFrame}
\alias{deskew_fcbFlowFrame}
\title{Corrects morphology based on scatter or uptake control for fcbFlowFrame}
\usage{
deskew_fcbFlowFrame(
  fcbFlowFrame,
  uptake = NULL,
  channel,
  method = "earth",
  predictors = c("fsc_a", "ssc_a"),
  subsample = 20000,
  ret.model = TRUE,
  verbose = FALSE,
  updateProgress = NULL,
  ...
)
}
\arguments{
\item{fcbFlowFrame}{a fcbFlowFrame object with barcoded flowframe and uptake flowframe, post compensation and preprocessing}

\item{channel}{The name (string) of the channel to be corrected, ie. the column name in fcbFlowFrame barcoded.ff exprs}

\item{method}{The name of the morphology correction model to use. Choose between earth, lm (linear model), or knijnenburg}

\item{predictors}{The vector of channel names to be used to build the regression model}

\item{subsample}{Integer, number of cells to sample (with replacement) for the morphology correction, defaults to 10,000.}

\item{ret.model}{Option to retain the model for deskewing}

\item{updateProgress}{used in reactive context (shiny) to return progress information to GUI}
}
\value{
a fcbFlowFrame with barcode slots added for selected channel corrected for predictors chosen
}
\description{
Corrects morphology based on scatter or uptake control for fcbFlowFrame
}
