% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_optimize.R
\name{em_optimize}
\alias{em_optimize}
\title{Defines populations on barcoded datasets}
\usage{
em_optimize(
  fcbFlowFrame,
  modelName = "VEE",
  subsample = 50000,
  verbose = TRUE,
  niter = 1,
  shrinkage = 0.05
)
}
\arguments{
\item{fcbFlowFrame}{a fcbFlowFrame object with barcoded flowframe and uptake flowframe post deskewing, clustering, and assignment}

\item{subsample}{number, number of cells to subsample during m step}

\item{ambiguitycut}{numeric from 0 to 1, threshhold below which to discard ambigious cells, eg: 0.02,
discards cells with more than 2% chance of originating from another population}
}
\value{
a fcbFlowFrame object with a barcode slot filled with deskewing, clustering, cell assignment as
a vector of integers from 0:ncol(probs), cells assigned a classification of 0 remained unassigned,
otherwise number corresponds to the barcoding level assignment of that cell
}
\description{
Defines populations on barcoded datasets
}
