% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcbFlowSet.R
\name{fcbFlowSet}
\alias{fcbFlowSet}
\title{Create an fcbFlowSet}
\usage{
fcbFlowSet(x)
}
\arguments{
\item{x}{A \code{flowSet} object.}
}
\value{
An object of class \code{fcbFlowSet}.
}
\description{
Constructor function to create an fcbFlowSet from a flowSet.
}
