% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRegressContrained.R
\name{generate_regressors}
\alias{generate_regressors}
\title{Used to build the regression model, not to be called directly by the used}
\usage{
generate_regressors(D, rm)
}
\value{
numeric vector of length nrow(fcb_df) representing the morphology corrected channel
}
\description{
Used to build the regression model, not to be called directly by the used
}
\seealso{
\code{\link{doRegressConstrained}}
}
