% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssignments.R
\name{getAssignments}
\alias{getAssignments}
\title{Extract assignments from an fcbFlowFrame or fcbFlowSet}
\usage{
getAssignments(x, platemap = NULL, simplify = FALSE)
}
\arguments{
\item{x}{An fcbFlowFrame or fcbFlowSet that has been assigned.}

\item{platemap}{Data frame, a lookup table from barcoding levels to well assignments (optional).}

\item{simplify}{Logical, return factor instead of list (not yet implemented).}
}
\value{
A list of factors, one for each barcoding channel that has been assigned.
}
\description{
Extract assignments from an fcbFlowFrame or fcbFlowSet
}
