% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssignments.R
\name{getAssignments}
\alias{getAssignments}
\title{Extracts assignments from fcbFlowFrame object}
\usage{
getAssignments(x, platemap = NULL, simplify = FALSE)
}
\arguments{
\item{platemap}{data.frame, a lookup table from barcoding levels to well assignments, (optional)}

\item{simplify}{logical, return factor instead of list}

\item{fcbFlowFrame}{a fcbFlowFrame object that has been assigned using assign_fcbFlowFrame}
}
\value{
a list of factors, one for each barcoding channel that has been assigned
}
\description{
Extracts assignments from fcbFlowFrame object
}
