% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jurkat-data.R
\docType{data}
\name{jurkatFCB}
\alias{jurkatFCB}
\title{48 Jurkat Samples Barcoded using Fluorescent Cell Barcoding}
\format{
An object of class \code{"flowFrame"}; see \code{\link[flowCore]{flowFrame-class}}.
}
\source{
Flow Repository link
}
\usage{
data(jurkatFCB)
}
\description{
Dataset consisting of 48 jurkat samples stained with combinations of
Pacific Orange (6 levels), Pacific Blue (8 levels), and Alexa Fluor 750
(1 level). Acquired as seperate FCS files and concatated together into a
single file, adding on row and col channels indicating the original row
and column for the data.
}
\examples{
data(jurkatFCB)
}
\references{
Reisman et al. (2020)
}
\keyword{datasets}
