% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology_corr_earth.R
\name{morphology_corr.earth}
\alias{morphology_corr.earth}
\title{Corrects morphology using multivariate adaptive regression splines as implemented in the earth pacakge}
\usage{
morphology_corr.earth(fcb, uptake, ret.model = FALSE, what = c("x",
  "x + se"), nfold = 1, ncross = 0, channel, predictors = c("FSC-A",
  "SSC-A"), subsample = 30000, updateProgress = NULL, ...)
}
\arguments{
\item{fcb}{the barcoded dataframe, post compensation and preprocessing}

\item{uptake}{Optional: a dataframe consisting of all cells barcoded with a single level of the barcoding dye}

\item{channel}{The name (string) of the channel to be corrected, ie. the column name in 'fcb_df'}

\item{predictors}{The vector of channel names to be used to build the regression model}

\item{subsample}{Integer, number of cells to sample (with replacement) for the morphology correction, defaults to 10,000.}

\item{updateProgress}{used in reactive context (shiny) to return progress information to GUI#'}
}
\value{
a tibble/data.frame with the selected channel corrected for fsc and ssc
}
\description{
Corrects morphology using multivariate adaptive regression splines as implemented in the earth pacakge
}
