% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology_corr_new.R
\name{morphology_corr}
\alias{morphology_corr}
\title{Corrects morphology based on scatter or uptake control.}
\usage{
morphology_corr(fcb, uptake = NULL, channel, method = c("earth",
  "knijnenburg", "lm"), predictors = c("FSC-A", "SSC-A"),
  subsample = 20000, exp_info = NULL, ret.model = FALSE,
  apply_scales = TRUE, updateProgress = NULL, ...)
}
\arguments{
\item{channel}{The name (string) of the channel to be corrected, ie. the column name in 'fcb_df'}

\item{subsample}{Integer, number of cells to sample (with replacement) for the morphology correction, defaults to 10,000.}

\item{updateProgress}{used in reactive context (shiny) to return progress information to GUI}

\item{fcb_df}{the barcoded dataframe, post compensation and preprocessing}

\item{bc_single_level}{Optional: a dataframe consisting of all cells barcoded with a single level of the barcoding dye}

\item{opt}{String, Morphology correction apporach, either "regression" (default) or "control dye"}

\item{trans}{String, transformation to apply to the barcoding channels, defaults to arcsinh, also log10}

\item{cofactor_bc1}{Numeirc, Cofactor used for the arcsinh transformation on the barcoding channels}

\item{cofactor_update}{Numeric, Cofactor used for the arcsinh transformation of the uptake channel}

\item{uptake_channel}{String, uptake channel if "control dye" option is used as morphology correction method}

\item{output}{what to return}
}
\value{
numeric vector of length nrow(fcb_df) representing the morphology corrected channel or a list containing the vector and the model used to correct it.
}
\description{
Corrects morphology based on scatter or uptake control.
}
