% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fcbFlowFrame-assignments.R
\name{plot.fcbFlowFrame}
\alias{plot.fcbFlowFrame}
\title{Plot debarcoding assignments}
\usage{
\method{plot}{fcbFlowFrame}(x, ..., plot = "assignments", seed = 1)
}
\arguments{
\item{x}{An fcbFlowFrame with completed assignments.}

\item{...}{Additional arguments (not used).}

\item{plot}{Character, plot type: "assignments" (default), "density", "chull", or "data".}

\item{seed}{Deprecated; no longer used.}
}
\value{
A ggplot2 object.
}
\description{
Plot debarcoding assignments
}
\examples{
data(jurkatFCB)
data(jurkatFCB_std)
fcb <- fcbFlowFrame(jurkatFCB)
fcb <- deskew_fcbFlowFrame(fcb,
    uptake = jurkatFCB_std,
    channel = "Pacific Blue-A",
    predictors = c("FSC-A", "SSC-A", "APC-H7-A")
)
fcb <- cluster_fcbFlowFrame(fcb, channel = "Pacific Blue-A", levels = 8)
fcb <- assign_fcbFlowFrame(fcb, channel = "Pacific Blue-A")
fcb <- deskew_fcbFlowFrame(fcb,
    uptake = jurkatFCB_std,
    channel = "Pacific Orange-A",
    predictors = c("FSC-A", "SSC-A", "APC-H7-A")
)
fcb <- cluster_fcbFlowFrame(fcb, channel = "Pacific Orange-A", levels = 6)
fcb <- assign_fcbFlowFrame(fcb, channel = "Pacific Orange-A")
plot(fcb)
}
