% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fcbFlowFrame-assignments.R
\name{plot.fcbflowframe}
\alias{plot.fcbflowframe}
\title{Plots assignemnts}
\usage{
plot.fcbflowframe(fcbFlowFrame, plot = "assignments", seed = 1)
}
\arguments{
\item{fcbFlowFrame}{a fcbFlowFrame object with barcoded flowframe and uptake flowframe post deskewing, clustering, and assignment}

\item{plot}{character, which plot to generate so far, only assignemnts have been implemented}
}
\value{
a fcbFlowFrame object with a barcode slot filled with deskewing, clustering, cell assignment as
a vector of integers from 0:ncol(probs), cells assigned a classification of 0 remained unassigned,
otherwise number corresponds to the barcoding level assignment of that cell
}
\description{
Plots assignemnts
}
