% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRegressContrained.R
\name{regression_model_matrix}
\alias{regression_model_matrix}
\title{Used to build the regression model, not to be called directly by the used}
\usage{
regression_model_matrix(
  nrow = 2,
  ncol = 8,
  vars = c(1, 1, 1, 2, 2, 2, 3, 3),
  powers = c(1, 0.5, 2, 1, 0.5, 2, 1, 0.5)
)
}
\value{
numeric vector of length nrow(fcb_df) representing the morphology corrected channel
}
\description{
Used to build the regression model, not to be called directly by the used
}
\examples{


}
\seealso{
\code{\link{selectDenseScatterArea}} \code{\link{doRegressConstrained}}
}
