% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDenseScatterArea.R
\name{selectDenseScatterArea}
\alias{selectDenseScatterArea}
\title{find the area which contains 95% of the data}
\usage{
selectDenseScatterArea(
  data,
  fsc_ssc = c(fsc = "FSC-A", ssc = "SSC-A"),
  subsample = 10000
)
}
\arguments{
\item{data}{text}

\item{subsample}{text}

\item{fsc}{text}

\item{ssc}{text}
}
\value{
some text
}
\description{
find the area which contains 95% of the data
}
\examples{
text
}
\seealso{
text
}
