% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_fcbFlowFrame.R
\name{split,fcbFlowFrame,list-method}
\alias{split,fcbFlowFrame,list-method}
\title{Split an fcbFlowFrame into an fcbFlowSet by assignments}
\usage{
\S4method{split}{fcbFlowFrame,list}(x, f, drop = FALSE, prefix = NULL, flowSet = TRUE, merge.na = TRUE, ...)
}
\arguments{
\item{x}{An fcbFlowFrame object with completed assignments.}

\item{f}{A list of assignment factors (from \code{\link{getAssignments}}).}

\item{drop}{Not used.}

\item{prefix}{Not used.}

\item{flowSet}{Logical, whether to return a flowSet (default TRUE).}

\item{merge.na}{Not used.}

\item{...}{Additional arguments (not used).}
}
\value{
An fcbFlowSet with each flowFrame named by assignment level,
  and pData columns for each barcoding channel.
}
\description{
Splits an fcbFlowFrame into separate flowFrames based on barcode
assignments, returned as an fcbFlowSet with updated pData.
}
\examples{
data(jurkatFCB)
data(jurkatFCB_std)
library(flowCore)
fcb <- fcbFlowFrame(jurkatFCB)
fcb <- deskew_fcbFlowFrame(fcb, uptake = jurkatFCB_std,
                           channel = "Pacific Blue-A",
                           predictors = c("FSC-A", "SSC-A", "APC-H7-A"))
fcb <- cluster_fcbFlowFrame(fcb, channel = "Pacific Blue-A", levels = 8)
fcb <- assign_fcbFlowFrame(fcb, channel = "Pacific Blue-A")
fcb <- deskew_fcbFlowFrame(fcb, uptake = jurkatFCB_std,
                           channel = "Pacific Orange-A",
                           predictors = c("FSC-A", "SSC-A", "APC-H7-A"))
fcb <- cluster_fcbFlowFrame(fcb, channel = "Pacific Orange-A", levels = 6)
fcb <- assign_fcbFlowFrame(fcb, channel = "Pacific Orange-A")
assignments <- getAssignments(fcb)
fcbfs <- split(fcb, assignments)
pData(fcbfs)
}
\seealso{
\code{\link{getAssignments}} to generate \code{f},
  \code{\link{apply_platemap}} to rename wells after splitting
}
