% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_fcbFlowFrame.R
\name{split,fcbFlowFrame,list-method}
\alias{split,fcbFlowFrame,list-method}
\title{Splits a fcbFlowFrame into a fcbFlowSet with updated pData}
\usage{
\S4method{split}{fcbFlowFrame,list}(x, f, drop = FALSE, prefix = NULL, flowSet = TRUE, merge.na = TRUE, ...)
}
\arguments{
\item{fcbFlowFrame}{a fcbFlowFrame object with barcoded flowframe and uptake flowframe post deskewing
and clustering (at least one barcodes slot filled)}
}
\value{
a fcbFlowSet with each flowFrame named by the assignments defined in the platemap
}
\description{
Splits a fcbFlowFrame into a fcbFlowSet with updated pData
}
