% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_fcbFlowFrame.R
\name{split,fcbFlowSet,list-method}
\alias{split,fcbFlowSet,list-method}
\title{Splits a fcbFlowSet into a flowSet with updated pData}
\usage{
\S4method{split}{fcbFlowSet,list}(
  x,
  f,
  assign0 = FALSE,
  drop = FALSE,
  prefix = NULL,
  flowSet = TRUE,
  merge.na = TRUE,
  ...
)
}
\arguments{
\item{fcbFlowFrame}{a fcbFlowFrame object with barcoded flowframe and uptake flowframe post deskewing
and clustering (at least one barcodes slot filled)}
}
\value{
a fcbFlowSet with each flowFrame named by the assignments defined in the platemap
}
\description{
Splits a fcbFlowSet into a flowSet with updated pData
}
