% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_fcbFlowFrame.R
\name{split,flowFrame,list-method}
\alias{split,flowFrame,list-method}
\title{Split a flowFrame by a list of factors}
\usage{
\S4method{split}{flowFrame,list}(x, f, drop = FALSE, prefix = NULL, flowSet = TRUE, merge.na = TRUE, ...)
}
\arguments{
\item{x}{A flowFrame object.}

\item{f}{A list of factors for splitting.}

\item{drop}{Not used.}

\item{prefix}{Not used.}

\item{flowSet}{Logical, whether to return a flowSet (default TRUE).}

\item{merge.na}{Not used.}

\item{...}{Additional arguments (not used).}
}
\value{
An fcbFlowSet with each flowFrame named by the combined factor levels.
}
\description{
Split a flowFrame by a list of factors
}
